net-p2p/amule/amule-9999-r1.ebuild

+++ net-p2p/amule/amule-9999.ebuild
@@ -1,4 +1,4 @@
-# Copyright 1999-2024 Gentoo Authors
+# Copyright 1999-2025 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
 EAPI=8
@@ -107,7 +107,7 @@ src_install() {
 		newinitd "${FILESDIR}"/amuleweb.initd amuleweb
 	fi
 
-	if use gui; then
+	if use gui && !use daemon; then
 		rm ${D}/usr/bin/amule
 	fi
 }
